machine-xilinx-default.inc: Default IMAGE_BOOT_FILES using function
Replace the definition of IMAGE_BOOT_FILES in machine-xilinx-board.inc
with the use of a function which automatically selects available images
that should be included. This includes the existing implementation of
`get_dtb_list` and replaces it with a function that covers all image
files including the existing default for KERNEL_IMAGETYPE(S) and
UBOOT_BINARY.

Also remove the use of `get_dtb_list` from individual machines which is
replaced by the default value.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>