machine-xilinx-board.inc: Remove include
This include has a single line which adds the virtual/bootloader to
EXTRA_IMAGEDEPENDS. Move this append into the individual machines and
drop the include. This makes using the meta-xilinx-bsp default
machine configuration much simpler for external users as well as making
the use of a bootloader explicit on a per machine basis.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>