{cl} Setup IMAGE_BOOT_FILES automatically
This series changes how the IMAGE_BOOT_FILES variable is set as well as
how the values in it are used by various parts of the meta-xilinx-bsp
layer. The IMAGE_BOOT_FILES value is now default populated in
machine-xilinx-default.inc to cover all machines in the meta-xilinx-bsp
layer.

The files populated in this variable now include all device-trees built
by device-tree.bb (if used) as well as all KERNEL_DEVICETREES defined by
the machine.conf. In order to get this working some changes were needed
to have u-boot-zynq-uenv capable of picking up the file names.

In order to achieve the automatic selection of device trees without
collecting all dtb files from the deploy directory (e.g.  undesired
duplicates of kernel built device trees), the device-tree recipe is
changed such that it deploys it's dtbs into a 'devicetree' subdirectory.

The default value of QB_DTB is also changed as it is now only populated
from IMAGE_BOOT_FILES.

Additionally this series removes machine-xilinx-board.inc, this is due
to it being reduced to a single EXTRA_IMAGEDEPENDS value.
---
Cc: Franz Forstmayr <f.forstmayr@gmail.com>
Cc: Manjukumar Harthikote Matha <manjukumar.harthikote-matha@xilinx.com>