device-tree: Consolidate device-tree recipe and append
With the introduction of devicetree.bbclass in oe-core/meta, remove the
implementation of device tree compilation from device-tree.bb keeping
the meta-xilinx specific information and licensing. Also consolidate the
device-tree.bb and device-tree.bbappend.

The only differences between the existing device-tree.bb implementation
and the devicetree.bbclass implementation is the device trees are
deployed in a "drivetree/" subdirectory within the deployed images
directory.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
---
devicetree.bbclass was added to oe-core/meta in the following commit:
 * http://git.openembedded.org/openembedded-core/commit/?id=3874c3859c3515fc8440fa9e776a02ad6ff3135c